# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wossolaty API Documentation",
        "description": "This is a wossolaty API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "haidershamil@gmail"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cash_details": {
            "get": {
                "tags": [
                    "Cash_Detail"
                ],
                "summary": "Returns API response of all Cash_Detail",
                "description": "A sample test of the API",
                "operationId": "getCash_DetailList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cash_Detail"
                ],
                "summary": "Store new Cash_Detail",
                "description": "Returns Cash_Detail data",
                "operationId": "storeCash_Detail",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/cash_details/{id}": {
            "get": {
                "tags": [
                    "Cash_Detail"
                ],
                "summary": "Get Cash_Detail information",
                "description": "Returns Cash_Detail data",
                "operationId": "getCash_DetailById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cash_Detail id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cash_Detail"
                ],
                "summary": "Update existing Cash_Detail",
                "description": "Returns updated Cash_Detail data",
                "operationId": "updateCash_Detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cash_Detail"
                ],
                "summary": "Delete existing Cash_Detail",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCash_Detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cash_Detail id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/cash_types": {
            "get": {
                "tags": [
                    "Cash_Type"
                ],
                "summary": "Returns API response of all Cash_Type",
                "description": "A sample test of the API",
                "operationId": "getCash_TypeList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cash_Type"
                ],
                "summary": "Store new Cash_Type",
                "description": "Returns Cash_Type data",
                "operationId": "storeCash_Type",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/cash_types/{id}": {
            "get": {
                "tags": [
                    "Cash_Type"
                ],
                "summary": "Get Cash_Type information",
                "description": "Returns Cash_Type data",
                "operationId": "getCash_TypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cash_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cash_Type"
                ],
                "summary": "Update existing Cash_Type",
                "description": "Returns updated Cash_Type data",
                "operationId": "updateCash_Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cash_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cash_Type"
                ],
                "summary": "Delete existing Cash_Type",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCash_Type",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Cash_Type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
...
}